Factor out location object encoding/decoding.
Also make a change to reduce the size of the serialized output.  If the start line and the end line are the same, then don't record the end line.

This is a small (but breaking) change to the serialization format.  However the docs are explicit about the format not being backwards compatible.