Set or unset .value if error is thrown.
Previous behaviour: after any uncaught error is thrown, the interpreter's .value property remains as whatever the last expression evaluated to.

New behaviour: In the case of an error in interpreted code, .value is set to the error thrown.  In the case of an error in the JS-Interpreter itself, .value is set to undefined.

Resolves #244